# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateSessionRequest:
    properties:
      name:
        example: my-session
        type: string
    required:
    - name
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: session not found
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      id:
        example: ABCD1234
        type: string
      message:
        example: operation completed
        type: string
      status:
        example: connected
        type: string
    type: object
...
  dto.PairPhoneRequest:
    properties:
      phone:
        example: "5511999999999"
        type: string
    required:
    - phone
    type: object
  dto.PairPhoneResponse:
    properties:
      code:
        example: ABCD-EFGH
        type: string
    type: object
...
  dto.QRResponse:
    properties:
      qr:
        example: data:image/png;base64,...
        type: string
      status:
        example: connecting
        type: string
    type: object
...
  dto.SessionResponse:
    properties:
      createdAt:
        example: "2025-11-29T14:18:15.324706Z"
        type: string
      deviceJid:
        example: 5511999999999@s.whatsapp.net
        type: string
      id:
        example: ce270f0c-c3d6-41ad-b481-1587f813c3b1
        type: string
      name:
        example: my-session
        type: string
      phone:
        example: "5511999999999"
        type: string
      status:
        example: connected
        type: string
      updatedAt:
        example: "2025-11-29T14:18:15.324706Z"
        type: string
    type: object
...
info:
  contact: {}
  description: WhatsApp API with Chatwoot Integration
  title: ZPWoot WhatsApp API
  version: 0.1.0
paths:
...
  /sessions:
    get:
      consumes:
      - application/json
      description: Get a list of all WhatsApp sessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SessionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Fetch all sessions
      tags:
      - sessions
    post:
      consumes:
      - application/json
      description: Create a new WhatsApp session with the given name
      parameters:
      - description: Session data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new session
      tags:
      - sessions
...
  /sessions/{sessionId}:
    delete:
      consumes:
      - application/json
      description: Delete an existing WhatsApp session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a session
      tags:
      - sessions
    get:
      consumes:
      - application/json
      description: Get information about a WhatsApp session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get session info
      tags:
      - sessions
...
  /sessions/{sessionId}/connect:
    post:
      consumes:
      - application/json
      description: Connect a WhatsApp session. If not authenticated, returns QR code
        endpoint
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Connect session
      tags:
      - sessions
...
  /sessions/{sessionId}/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnect from WhatsApp but keep credentials (can auto-reconnect)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Disconnect session
      tags:
      - sessions
...
  /sessions/{sessionId}/logout:
    post:
      consumes:
      - application/json
      description: Logout from WhatsApp and clear credentials (requires new QR scan
        to reconnect)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout session
      tags:
      - sessions
...
  /sessions/{sessionId}/pair/phone:
    post:
      consumes:
      - application/json
      description: Pair WhatsApp session using phone number instead of QR code
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PairPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PairPhoneResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Pair using phone number
      tags:
      - sessions
...
  /sessions/{sessionId}/qr:
    get:
      consumes:
      - application/json
      description: Get QR code for WhatsApp authentication
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - default: json
        description: Response format (json or image)
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.QRResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get QR code
      tags:
      - sessions
...
  /sessions/{sessionId}/restart:
    post:
      consumes:
      - application/json
      description: Disconnect and reconnect a WhatsApp session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Restart session
      tags:
      - sessions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"
...